Lots of stuff

+++ src/components/ItemList.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 import Item from "../components/Item";
 import styled, { ThemeProvider } from "styled-components";
 import { theme } from "../theme";
+import { getItemById } from "../utils";
 
 import { connect } from "react-redux";
 import { deleteItem, refileItem } from "../actions";
@@ -14,22 +15,44 @@ const NoItemText = styled.p`
   padding-left: 10px;
 `;
 
+const Container = styled.div`
+  margin: 10px 0px;
+`;
+
 function ItemList(props) {
   return (
     <ThemeProvider theme={theme}>
-      {props.items.map((item, index) => {
-        return (
-          <Item
-            {...item}
-            key={item.id}
-            prevItemId={index > 0 ? props.items[index - 1].id : null}
-            onDelete={props.deleteItem}
-            onRefile={props.refileItem}
-          />
-        );
-      })}
-
-      {props.items.length == 0 && <NoItemText>No items</NoItemText>}
+      <Container>
+        {props.items.map((item, index) => {
+          const isSubtask = item.parentId != null;
+          if (!props.showSubtasks && isSubtask) return;
+          return (
+            <div>
+              <Item
+                {...item}
+                key={item.id}
+                onDelete={props.deleteItem}
+                onRefile={props.refileItem}
+                noIndentation={props.noIndentation}
+              />
+              {item.children &&
+                item.children.map(c => {
+                  const childItem = getItemById(c, props.items);
+                  return (
+                    <Item
+                      {...childItem}
+                      key={c}
+                      onDelete={props.deleteItem}
+                      onRefile={props.refileItem}
+                      noIndentation={props.noIndentation}
+                    />
+                  );
+                })}
+            </div>
+          );
+        })}
+        {props.items.length == 0 && <NoItemText>No items</NoItemText>}
+      </Container>
     </ThemeProvider>
   );
 }
